Cnms dev

+++ pycroscopy/analysis/utils/be_sho.py
@@ -5,6 +5,7 @@ Created on Mon Sep 28 11:35:57 2015
 @author: Anton Ievlev
 """
 
+from __future__ import division, print_function, absolute_import
 import numpy as np
 from numpy import exp, abs, sqrt, sum, real, imag, arctan2, append
 
@@ -122,7 +123,7 @@ def SHOfastGuess(w_vec, resp_vec, qual_factor=200):
         SHO fit parameters arranged as [amplitude, frequency, quality factor, phase]
     """
     amp_vec = abs(resp_vec)
-    i_max = len(resp_vec)/2
+    i_max = int(len(resp_vec)/2)
     return np.array([np.mean(amp_vec) / qual_factor, w_vec[i_max], qual_factor, np.angle(resp_vec[i_max])])
 
 def SHOlowerBound(w_vec):
